[libcalamares] Merge Tests.h into Tests.cpp
- no real need for separate file in the utils/ tests
